handle clearing the db and initialize cases

+++ medallion/backends/mongodb_backend.py
@@ -61,16 +61,19 @@ class MongoBackend(Backend):
 
     def __init__(self, **kwargs):
         try:
-            super(MongoBackend, self).__init__(**kwargs)
+
             self.pages = {}
             self.client = MongoClient(kwargs.get("uri"))
 
-            if kwargs.get("filename"):
-                log.info("Initializing Mongo DB backend using " + kwargs.get("filename"))
+            if kwargs.get("clear_db"):
                 self._clear_db()
-                self.initialize_mongodb(kwargs.get("filename"))
-
-            self.object_manifest_check()
+                if kwargs.get("filename"):
+                    log.info("Initializing Mongo DB backend using " + kwargs.get("filename"))
+                    self.initialize_mongodb(kwargs.get("filename"))
+                    self.object_manifest_check()
+            else:
+                self.object_manifest_check()
+            super(MongoBackend, self).__init__(**kwargs)
 
         except ConnectionFailure:
             log.error("Unable to establish a connection to MongoDB server {}".format(kwargs.get("uri")))
@@ -125,37 +128,38 @@ class MongoBackend(Backend):
     def _pop_old_statuses(self):
         if "discovery_database" in self.client.list_database_names():
             api_roots = self._get_all_api_roots()
-            status_retention_in_milliseconds = self.status_retention * 1000
-            for ar in api_roots:
-                statuses_of_api_root = self._get_api_root_statuses(ar)
-                result = statuses_of_api_root.aggregate([
-                        {
-                            "$project": {
-                                "id": 1,
-                                "date_difference": {
-                                    "$subtract": [
-                                        "$$NOW",
-                                        {
-                                            "$dateFromString": {
-                                                "dateString": "$request_timestamp"
+            if api_roots:
+                status_retention_in_milliseconds = self.status_retention * 1000
+                for ar in api_roots:
+                    statuses_of_api_root = self._get_api_root_statuses(ar)
+                    result = statuses_of_api_root.aggregate([
+                            {
+                                "$project": {
+                                    "id": 1,
+                                    "date_difference": {
+                                        "$subtract": [
+                                            "$$NOW",
+                                            {
+                                                "$dateFromString": {
+                                                    "dateString": "$request_timestamp"
+                                                }
                                             }
-                                        }
-                                    ]
-                                },
-                            }
-                        },
-                        {
-                            "$match": {
-                                "date_difference": {
-                                    "$gt": status_retention_in_milliseconds
+                                        ]
+                                    },
+                                }
+                            },
+                            {
+                                "$match": {
+                                    "date_difference": {
+                                        "$gt": status_retention_in_milliseconds
+                                    }
                                 }
                             }
-                        }
-                    ]
-                )
-                for doc in result:
-                    log.info("Status {} was deleted from {} because it was older than the status retention time".format(doc["id"], ar))
-                    statuses_of_api_root.delete_one({"_id": doc["_id"]})
+                        ]
+                    )
+                    for doc in result:
+                        log.info("Status {} was deleted from {} because it was older than the status retention time".format(doc["id"], ar))
+                        statuses_of_api_root.delete_one({"_id": doc["_id"]})
 
     def _get_object_manifest(self, api_root, collection_id, filter_args, allowed_filters, limit, internal=False):
         api_root_db = self.client[api_root]
@@ -233,7 +237,8 @@ class MongoBackend(Backend):
         discovery_info = discovery_db["discovery_information"]
         info = discovery_info.find_one()
         # this could also be done via the mongo db finds_one query
-        info.pop("_id")
+        if info:
+            info.pop("_id")
         return info
 
     @catch_mongodb_error
@@ -488,17 +493,15 @@ class MongoBackend(Backend):
 
     def initialize_mongodb(self, filename):
         self.load_data_from_file(filename)
-        self.client.drop_database("discovery_database")
         if "/discovery" in self.json_data:
             db = self.client["discovery_database"]
             db["discovery_information"].insert_one(self.json_data["/discovery"])
         else:
             raise InitializationError("No discovery information provided when initializing the Mongo DB")
+        api_root_info_db = db["api_root_info"]
         for api_root_name, api_root_data in self.json_data.items():
             if api_root_name == "/discovery":
                 continue
-            else:
-                api_root_info_db = db["api_root_info"]
             url = list(filter(lambda a: api_root_name in a, self.json_data["/discovery"]["api_roots"]))[0]
             api_root_data["information"]["_url"] = url
             api_root_data["information"]["_name"] = api_root_name
@@ -544,10 +547,15 @@ class MongoBackend(Backend):
 
     def _clear_db(self):
         if "discovery_database" in self.client.list_database_names():
+            log.info("Clearing database")
             discovery_db = self.client["discovery_database"]
             api_root_info = discovery_db["api_root_info"]
             for api_info in api_root_info.find({}):
                 self.client.drop_database(api_info["_name"])
             self.client.drop_database("discovery_database")
-            db = self.client["discovery_database"]
-            return db
+        # db with empty tables
+        log.info("Creating empty database")
+        discovery_db = self.client.get_database("discovery_database")
+        discovery_db.create_collection("discovery_information")
+        discovery_db.create_collection("api_root_info")
+        return discovery_db
